Example of Biomarker Grouping Data Frame for All Nightingale Biomarkers

A data frame containing all the Nightingale Health Biomarkers' in a machine readable name format along with a custom grouping, a 2-page and a 2-column specification. This data frame is used by plot_all_NG_biomarkers.

df_grouping_all_NG_biomarkers

Format

A data frame (tibble) with 228 rows and 4 columns:

machine_readable_name

Biomarker machine readable name, i.e. the one delivered with the csv data format.

group_custom

An example of a slightly different grouping than the one provided in df_NG_biomarker_metadata.

column

An integer indicating the column number for plotting in a certain layout, see plot_all_NG_biomarkers.

page

An integer indicating the page number for plotting in a certain layout, see plot_all_NG_biomarkers.

See also