Odds Ratios of Blood Biomarkers with Incident Type 2 Diabetes

A data frame containing log odds ratios of the Nightingale blood biomarkers with risk for future type 2 diabetes.

df_logodds_associations

Format

A data frame (tibble) with 228 rows and 5 columns:

name

Biomarker abbreviation.

study

Short description of the association type, here the cohort name.

beta

Log odds for incident type 2 diabetes.

se

Standard error.

pvalue

P-value.

n

Sample size.

Source

These data are taken from the Supplementary material of A. V. Ahola-Olli et al. (2019). https://www.biorxiv.org/content/early/2019/01/08/513648

Details

"Values are odds ratios (95% confidence intervals) per 1-SD log-transformed metabolite concentration. Odds ratios were adjusted for sex, baseline age, BMI, and fasting glucose. YFS, Cardiovascular risk in Young Finns Study; NFBC, Northern Finland Birth Cohort." The values are shown in Figure S3 of A. V. Ahola-Olli et al. (2019): https://www.biorxiv.org/content/early/2019/01/08/513648