ggforestplot: Forestplot Visualizations.

ggforestplot is a package for plotting measures of effect, e.g. linear associations or log and hazard ratios, with their confidence interval in a forestplot layout. It's main plotting function is forestplot().